Get genomic ranges based on specified features and flanking regions.

This function takes a GRanges object and extracts or creates new genomic ranges based on specified start and end features (TSS, TES, or Exon) and their flanking regions. It allows for specifying exon numbers and boundaries for precise range definition.

getFeature(
  object,
  start_feature = "TSS",
  start_flank = 0,
  start_exon = NULL,
  start_exon_boundary = NULL,
  start_direction = "up",
  end_feature = "TES",
  end_flank = 0,
  end_direction = "down",
  end_exon = NULL,
  end_exon_boundary = NULL,
  return = "all"
)

Arguments

object

A GRanges object containing genomic features (e.g., genes with exons). This object must have metadata columns including 'blockStarts', 'blockSizes', and 'blockCount' if 'Exon' is specified as a feature.

start_feature

A character string specifying the starting feature. Must be one of "TSS" (Transcription Start Site), "TES" (Transcription End Site), or "Exon". Default is "TSS".

start_flank

An integer indicating the number of bases to flank the start_feature. Default is 0.

start_exon

An integer specifying the exon number to use if start_feature is "Exon".

start_exon_boundary

A character string specifying the exon boundary to use if start_feature is "Exon". Must be either "5prime" or "3prime".

start_direction

A character string indicating the direction of the flank relative to the start_feature. Must be either "up" (upstream) or "down" (downstream). Default is "up".

end_feature

A character string specifying the ending feature. Must be one of "TSS", "TES", or "Exon". Default is "TES".

end_flank

An integer indicating the number of bases to flank the end_feature. Default is 0.

end_direction

A character string indicating the direction of the flank relative to the end_feature. Must be either "up" (upstream) or "down" (downstream). Default is "down".

end_exon

An integer specifying the exon number to use if end_feature is "Exon".

end_exon_boundary

A character string specifying the exon boundary to use if end_feature is "Exon". Must be either "5prime" or "3prime".

return

A character string specifying which ranges to return. Must be one of "all" (both plus and minus strand ranges), "plus" (only plus strand ranges), or "minus" (only minus strand ranges). Default is "all".

Value

A GRanges object containing the extracted or created genomic ranges.

Examples

# Create a dummy GRanges object
library(GenomicRanges)
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
genes <- GRanges(
  seqnames = "chr1",
  ranges = IRanges(start = c(100, 2000), end = c(1000, 3000)),
  strand = c("+", "-"),
  blockStarts = c("0,200,500", "0,300"),
  blockSizes = c("100,300,500", "200,500"),
  blockCount = c(3, 2)
)

# Get TSS regions with 100bp upstream flank
tss_regions <- getFeature(genes, start_feature = "TSS", start_flank = 100)
tss_regions
#> GRanges object with 2 ranges and 3 metadata columns:
#>       seqnames    ranges strand | blockStarts  blockSizes blockCount
#>          <Rle> <IRanges>  <Rle> | <character> <character>  <numeric>
#>   [1]     chr1 2000-3100      - |       0,300     200,500          2
#>   [2]     chr1    0-1000      + |   0,200,500 100,300,500          3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Get regions from TSS to TES
gene_bodies <- getFeature(genes, start_feature = "TSS", end_feature = "TES")
gene_bodies
#> GRanges object with 2 ranges and 3 metadata columns:
#>       seqnames    ranges strand | blockStarts  blockSizes blockCount
#>          <Rle> <IRanges>  <Rle> | <character> <character>  <numeric>
#>   [1]     chr1 2000-3000      - |       0,300     200,500          2
#>   [2]     chr1  100-1000      + |   0,200,500 100,300,500          3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Get regions around the 2nd exon start
exon2_start <- getFeature(genes, start_feature = "Exon", start_exon = 2,
                           start_exon_boundary = "5prime", start_flank = 50)
exon2_start
#> GRanges object with 2 ranges and 3 metadata columns:
#>       seqnames    ranges strand | blockStarts  blockSizes blockCount
#>          <Rle> <IRanges>  <Rle> | <character> <character>  <numeric>
#>   [1]     chr1 2000-2250      - |       0,300     200,500          2
#>   [2]     chr1  250-1000      + |   0,200,500 100,300,500          3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

# Get regions spanning the 1st to the 2nd exon
exon1_to_2 <- getFeature(genes, start_feature = "Exon", start_exon = 1,
                          start_exon_boundary = "5prime",
                          end_feature = "Exon", end_exon = 2,
                          end_exon_boundary = "3prime")
exon1_to_2
#> GRanges object with 2 ranges and 3 metadata columns:
#>       seqnames    ranges strand | blockStarts  blockSizes blockCount
#>          <Rle> <IRanges>  <Rle> | <character> <character>  <numeric>
#>   [1]     chr1 2000-2800      - |       0,300     200,500          2
#>   [2]     chr1   100-600      + |   0,200,500 100,300,500          3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths